feat: support decimal sleep quality; drop redundant ios history tab
Apple Health-style scores can arrive on a 0-100 scale; storing quality
as a float lets a synced value map to the 1-10 scale (e.g. 9.7/10) while
the manual slider stays integer 1-10.

- Server: quality column int -> real, validation drops the integer
  constraint, regenerated OpenAPI spec + Kotlin DTOs.
- Web + iOS + Android: display one decimal only when fractional.
- iOS/Android cache keep the denormalized quality column integer; the
  real value round-trips through the serialized entry.
- iOS: remove the sleep chart 'History' range tab (identical to 90d),
  leaving 7d/30d/90d.

+++ mobile/iosApp/Bissbilanz/Models/Sleep.swift
@@ -5,7 +5,7 @@ struct SleepEntry: Codable, Identifiable {
     let userId: String
     let entryDate: String
     let durationMinutes: Int
-    let quality: Int
+    let quality: Double
     let bedtime: String?
     let wakeTime: String?
     let wakeUps: Int?
@@ -22,7 +22,7 @@ struct SleepEntry: Codable, Identifiable {
 
 struct SleepCreate: Codable {
     let durationMinutes: Int
-    let quality: Int
+    let quality: Double
     let entryDate: String
     var bedtime: String?
     var wakeTime: String?
@@ -32,7 +32,7 @@ struct SleepCreate: Codable {
 
 struct SleepUpdate: Codable {
     var durationMinutes: Int?
-    var quality: Int?
+    var quality: Double?
     var entryDate: String?
     var bedtime: String?
     var wakeTime: String?
